Hypothesis test
With the availability of ample computing power and statistical computing tools such as R, we no longer have to live with the limitations. Sampling and creating sampling distributions are simple enough tasks, and we can construct our own sampling distribution for the test that we are interested in conducting. This approach of creating your own empirical distributions is termed permutation (or randomization) testing.
Key advantages of permutation testing include:
1. Fewer assumptions – for example, you don’t have to have large sample sizes
2. Generality – can be conducted for a wide range of statistics. For example, ratio of means (between males and females) testing for which there is no traditional means of testing.
3. Greater accuracy – in many cases can be more accurate than classical methods
4. Promote understanding – more intuitive and concrete.

The general approach to conduct permutation testing is:
1. Choose a statistic of interest (eg. Mean, Percentile, Correlation, etc)
2. Generate a hypothesis that enables you to describe the (hypothesized) population
3. Draw samples from the population and construct the sampling distribution
4. Locate the observed statistic on this distribution to draw your inference